attacks/xss/requirements.txt
# Core dependencies
Flask==3.1.0

# Database
Flask-SQLAlchemy==3.1.1

# XSS Attack Tools
playwright==1.51.0

python-dotenv==1.1.0